ai-agent/requirements.txt
openai
fastapi
uvicorn
python-dotenv
pydantic
requests
